Move implementation outside of mat module
